pyproject.toml
[project]
name = "deepanal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "catppuccin[rich]>=2.5.0",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.10",
    "langchain-ollama>=0.3.7",
    "langgraph>=0.6.6",
    "rich>=14.1.0",
    "tavily-python>=0.7.11",
]
